Create KeyCracker class and methods
