+++ lib/crack.rb
@@ -4,6 +4,7 @@ require_relative 'cracker'
 require_relative 'dateable'
 require_relative 'enigma'
 require_relative 'key'
+require_relative 'key_cracker'
 require_relative 'offset'
 require_relative 'positionable'
 require_relative 'shiftable'
